Upgrade dependencies packages and python 3.12
