+++ deployment/docker/rockylinux/Dockerfile
@@ -6,7 +6,9 @@ RUN dnf update -y
 RUN dnf groupinstall "Development Tools" -y
 RUN  dnf install libpq-devel -y
 RUN dnf install python3-pip -y
-RUN dnf install -y   python3-devel.x86_64
+RUN dnf install "python3.12-devel.x86_64" -y
+RUN dnf install  python3.12  -y
+RUN dnf install -y  python3.12-pip -y
 RUN dnf clean all && rm -rf /var/cache/yum
 RUN mkdir /searchengine
 ADD  deployment/docker/rockylinux/start_gunicorn_serch_engine.sh /searchengine
@@ -17,7 +19,7 @@ RUN mkdir /etc/searchengine
 RUN mkdir /etc/searchengine/chachedata
 RUN mkdir /etc/searchengine/logs
 WORKDIR /searchengine
-RUN pip3 install -r  requirements.txt
-RUN pip3 install gunicorn
+RUN pip3.12 install -r  requirements.txt
+RUN pip3.12 install gunicorn
 EXPOSE 5577
 ENTRYPOINT ["bash", "run_app.sh"]
