+++ requirements.txt
@@ -1,35 +1,39 @@
-amqp==5.0.6
-billiard==3.6.4.0
-cffi==1.14.6
-click==7.1.2
-click-didyoumean==0.0.3
-click-plugins==1.1.1
-click-repl==0.2.0
-cryptography==3.4.7
+attrs==25.3.0
+babel==2.17.0
+blinker==1.9.0
+certifi==2025.6.15
+charset-normalizer==3.4.2
+click==8.2.1
+elastic-transport==8.17.1
 elasticsearch==7.15.1
-Flask==1.1.4
+flasgger==0.9.7.1
+Flask==3.1.1
+flask-babel==4.0.0
 Flask-Script==2.0.6
-flasgger==0.9.5
-greenlet==1.1.0
-itsdangerous==1.1.0
-Jinja2==2.11.3
-kombu==5.1.0
-MarkupSafe==2.0.1
-numexpr==2.7.3
-numpy==1.21.2
-pandas==1.3.2
-pickle-mixin==1.0.2
-prompt-toolkit==3.0.20
-psycopg2==2.9.1
-pycparser==2.20
-python-dateutil==2.8.2
-pytz==2021.1
-PyYAML==5.4.1
-requests==2.26.0
-six==1.16.0
+greenlet==3.2.3
+idna==3.10
+itsdangerous==2.2.0
+Jinja2==3.1.6
+jsonschema==4.24.0
+jsonschema-specifications==2025.4.1
+MarkupSafe==3.0.2
+mistune==3.1.3
+numpy==2.3.1
+packaging==25.0
+pandas==2.3.0
+psycopg2-binary==2.9.10
+python-dateutil==2.9.0.post0
+pytz==2025.2
+PyYAML==6.0.2
+referencing==0.36.2
+requests==2.32.4
+rpds-py==0.25.1
+six==1.17.0
 SQLAlchemy==1.4.21
-vine==5.0.0
-wcwidth==0.2.5
-Werkzeug==1.0.1
-bump2version==1.0.1
+style==1.1.0
+typing_extensions==4.14.0
+tzdata==2025.2
+update==0.0.1
+urllib3==1.26.20
 watchdog==6.0.0
+Werkzeug==3.1.3
